feat(updated): add useSearch hook
