+++ src/screens/Search/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useSearch} from './useSearch'
